Add release count metric
As we want to know the number of relases performed by the Helm
Operator, a counter metric (`releases_performed_total`) was added
to keep track of this number.
Both installations and upgrades increase the counter.

*Examples of how to use the metric in a PromQL query:*
Using the metric "as is"
(`flux_helm_operator_releases_performed_total`) returns the total
number of releases since the Helm Operator was started.

To get the total number of releases performed during the last
7 days:
`sum(max_over_time(flux_helm_operator_releases_performed_total[7d]))`

Co-authored-by: larsroger <larsroger@users.noreply.github.com>