Release: remove failed initial release
To make Helm v3 a first class citizen, it would be better to mimic
the `--atomic` behaviour from Helm v3 in the v2 client code, so that
the release is removed automagically if the installation fails.

This is however outside the scope of the PR this commit is part of.