Release: restore Prometheus metric
The previous metric kept track of the time it would take to perform
a Helm action, and some other actions around the synchronization of
the `HelmRelease` resource to the cluster, but would not give a real
insight in the total amount of time it would take to process a
`HelmRelease` resource.

This commit slightly tweaks the metric, so that it now measures the
total amount of time it takes to synchronize a single `HelmRelease`
resource to the cluster, and drops labels that due to this change
have become irrelevant.