Release: only set `atomic` option during install
If the atomic option is set during upgrades, Helm v3 will automatically
rollback the release if it fails, which is not something we are after
as this can be dangerous for e.g. charts that make use of `StatefulSet`
resources.

During installation however we _do_ want to use this flag, as this
automatically removes a failed release for us, so that we do not have
to perform or implement this logic ourselves.