helm/v3: always YAML parse returned release values
This commit adds a `configToGenericValues` method to the v3 client
layer that forces the returned release values to be (re)parsed as
YAML so that the value types of the nested map always equal to what
they would be when returned from storage.

The main reason we have to do this is because a dry-run release seems
to skip this step and return the values directly from memory, causing
difficulties during the comparison we make to see if we actually have
to make a release, as e.g. float values will be strings when returned
from storage (current release), but in the values of the dry-run
release they will still be floats.