release: declare `AntecedentAnnotation`
This commit drops the dependency on the `cluster/kubernetes` package
from Flux by moving the `AntecedentAnnotation` to our own code base.

The main reason to do this now is because we want to be able to define
our own Kubernetes package versions, something that is not possible
without this change due to `cluster/kubernetes` (indirectly) depending
on various `k8s.io/*` packages.