operator: add `--helm-repository-import` flag
With this flag it is possible to import repositories from an index
file to a specific Helm client version.

Example:

`--helm-repository-import=v2:/var/v2.yaml,v3:/var/v3.yaml` causes
the operator to import the `v2.yaml` repositories in the Helm v2
client, and the `v3.yaml` repositories in the Helm v3 client.