helm: tidy operation options and remove validators
The validators were no longer beneficial to the user as they are passed
regardless of the version targeted by the release, due to the refactor
of the `release` package.

The amount of files was also getting out of hand and making it
difficult to navigate through the package. Therefor all operation
options have been bundled in `options.go`, and a descriptive comment
has been added about how they should be implemented.