release: run dry-run upgrade instead of install
Due to the server side validation Helm v3 introduced, we were forced
into running it with the `ClientOnly` option enabled as else it would
complain about cluster wide resources already existing in the cluster.

The goal of the dry-run is to determine if it changes anything to
what we are currently running, and thus if we should proceed with
the release. This commit changes the dry-run from install to upgrade,
with the configured release name instead of the UID of the
`v1.HelmRelease`.

For releases targeting Helm v3 this also ensures that we only perform
an upgrade if there are no validation errors which restores the rollback
logic. As before this change validation errors would not result in an
actual release but we would still attempt to perform a rollback,
resulting in either a rollback to an incorrect version, or worse, an
accidental uninstall.