fix: authentication issue in e2e test
In the `15_upgrade` tests when it was doing a git clone it would
report `Too many authentication failures` if there are a number
of different ssh keys on the system. As the tests use a generated
key pair we can tell ssh to only use the supplied identity file and
not try all found keys.

Relates to: #179