e2e: use proper helm.versions value to set Helm version
The `HELM_VERSION` environmental variable is only used when the
flag has not been modified. Due to the default `helm.versions`
value in the Helm chart's `values.yaml`, the flag is always
modified when the Helm operator is installed using Helm and the
environmental value is thus ignored.

This commit sets the `helm.versions` flag when installing the
operator for end-to-end tests instead of the `HELM_VERSION`, so
that the correct Helm version is targeted for e.g. Helm v3 specific
tests.