helm/v2: if present, use repository alias to pull a chart
This is a workaround for some problems that exist in Helm while
attempting to fetch a chart with an absolute URL, while it also
makes things a bit more efficient.

The problem is that the authentication credentials for an URL are
looked up by Helm based on the URLs of the index file that is
present for the repository. This itself is not a problem as long
as the chart repository is configured to list absolute URLs in
their index, some repositories however are configured to list
relative URLs, which will make the credential lookup fail.