release: rely on Helm storage for rollbacks
This commit removes our own state keeping around rollbacks except
for the condition we set marking a rollback has happened, but
instead relies on comparison of a dry-run with the latest failed
release. If this results in a diff, a new upgrade will be attempted.

The reason we move to this approach is that it is not possible to
take all variables safely into account and determine if we should
perform a new upgrade attempt based on our own bookkeeping, with
a long list of edge-case bugs as a result.

Relying on the real truth is always better than a reflection of it,
which in the end may not be the actual truth at all.