helm: do not include non-template files in chart data
As the non-template files include the `requirements.lock` file,
which in case of a chart with dependencies differs every time
a dry-run happens due to the generation timestamp it includes,
resulting in spurious upgrades.

Other files included there are non-trivial for the detection of
changes as their effects are either captured in other parts of
the chart metadata, or do not matter at all for the release
itself, like the contents of a `README.md`.