release: introduce reuse-values functionality
This commit is a fix of a regression bug, or faulty description of a
feature.

In a previous version of the Helm Operator the `ResetValues` option
was introduced to make Helm reset the values to the chart defaults
during an upgrade. In a later version this behaviour changed, and
values were reset by default (as expected given a HelmRelease
resource should be the sole truth).

This commit adds the possibility to reuse values from previous
releases by explicitly setting `.spec.resetValues` to `false`. This
will make an upgrade made by the operator behave as if the
`--reuse-values` flag would be set on an `helm upgrade` command.

Removing this defintion from the HelmRelease, or setting
`.resetValues` to true will make the operator change back to its
default behaviour during subsquent upgrades made for the release.