Update role, rolebinding, (cluster...) to v1 apis
The cluster rbac and rbac roles in the helm chart templates can still
reference the old v1beta1 rbac.authorization.k8s.io API, which must be
removed in preparation for K8s 1.22.

The CRD changes already meant that Kubernetes 1.16+ was required in the
(next) newest version of Helm Operator.

This change requires K8s 1.8+ so, while we can mention it separately in
the release notes, it does not add any additional functional requirement
for the users, as they must already upgrade to this version or greater.

Signed-off-by: Kingdon Barrett <yebyen@gmail.com>